desktop-app/client/requirements.txt
absl-py==2.3.1
astor==0.8.1
astunparse==1.6.3
attrs==25.3.0
autocommand==2.2.2
bcrypt==4.3.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.8.2
build==1.3.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.3
cryptography==45.0.5
cycler==0.12.1
dbus-python==1.4.0
distro==1.9.0
evdev==1.9.2
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.59.0
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
gps==3.26.1
grpcio==1.74.0
grpcio-tools==1.74.0
h5py==3.14.0
html5lib==1.1
httplib2==0.22.0
hyprshade==4.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.11.1
kiwisolver==1.4.7
lark==1.2.2
lazy_loader==0.4
lensfun==0.3.4
libtorrent==2.0.11
lxml==6.0.0
Mako==1.3.10.dev0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
meson==1.8.3
ml_dtypes==0.5.1
moddb==0.14.0
more-itertools==10.7.0
networkx==3.5
nibabel==5.3.2
nilearn==0.12.0
numpy==2.3.2
oauth2client==4.1.3
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.6
poetry-core==2.1.3
pooch==1.8.2
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==3.0.0
pycairo==1.28.0
pycparser==2.22
pydicom==3.0.1
pydot==3.0.1
pygdbmi==0.11.0.0
Pygments==2.19.2
PyGObject==3.52.3
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0
pytz==2025.2
pywalfox==2.7.4
pyxdg==0.28
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==1.3.1
rich==14.1.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
screeninfo==0.8.1
sentry-sdk==2.34.1
setuptools==80.9.0
six==1.17.0
smbus==1.1
soupsieve==2.6
standard-cgi==3.13.0
TBB==0.2
tensorboard==2.20.0
tensorboard_data_server==0.8.0a0
tensorboard_plugin_wit==1.8.1
tensorflow_cpu==2.19.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.6.11
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
waypaper==2.6
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.16.0
zipp==3.21.0
